requirements.txt
SpeechRecognition
pyttsx3
requests
pytz
openai
sympy
python-dotenv
flask
flask-cors
# langchain
# transformers
# accelerate
# sentence-transformers
# faiss-cpu